# back/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Oasix property management system API server.",
        "title": "Oasix API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/core/locales/{lang}/{ns}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Keycloak": []
                    },
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get locales",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "core"
                ],
                "summary": "Get locales",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language",
                        "name": "lang",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "ns",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/core/properties-enums": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Keycloak": []
                    },
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get properties enums",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "core"
                ],
                "summary": "Get properties enums",
                "parameters": [
                    {
                        "description": "Filter parameters",
                        "name": "filter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_repository.PropertiesEnumFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.PropertiesEnumsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/core/properties-types": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Keycloak": []
                    },
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get paginated list of properties types with filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "core"
                ],
                "summary": "Get paginated list of properties types",
                "parameters": [
                    {
                        "description": "Filter parameters",
                        "name": "filter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_repository.PropertiesTypeFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.PropertiesTypesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/core/properties-types/{id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Keycloak": []
                    },
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get properties type by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "core"
                ],
                "summary": "Get properties type by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Properties type id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Filter parameters",
                        "name": "filter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.DefaultRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.PropertiesTypeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/parier-server_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handlers.PropertiesEnumResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "property": {
                    "type": "string"
                },
                "property_type": {
                    "$ref": "#/definitions/parier-server_internal_models.PropertyType"
                },
                "value": {},
                "value_id": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.PropertiesEnumsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_handlers.PropertiesEnumResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_handlers.PropertiesTypeResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "place": {
                    "$ref": "#/definitions/parier-server_internal_models.PropertyPlace"
                },
                "type": {
                    "$ref": "#/definitions/parier-server_internal_models.PropertyType"
                }
            }
        },
        "internal_handlers.PropertiesTypesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_handlers.PropertiesTypeResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "parier-server_internal_models.DefaultRequest": {
            "type": "object",
            "properties": {
                "language": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/parier-server_internal_models.User"
                }
            }
        },
        "parier-server_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "parier-server_internal_models.PropertyPlace": {
            "type": "string",
            "enum": [
                "AGENT",
                "UNIT",
                "USER"
            ],
            "x-enum-varnames": [
                "PropertyPlaceAgent",
                "PropertyPlaceUnit",
                "PropertyPlaceUser"
            ]
        },
        "parier-server_internal_models.PropertyType": {
            "type": "string",
            "enum": [
                "DATE",
                "DECIMAL",
                "INTEGER",
                "LOCALIZATION",
                "BOOLEAN",
                "TEXT",
                "JSONOBJECT",
                "JSONARRAY",
                "MEDIA",
                "ENUM"
            ],
            "x-enum-varnames": [
                "PropertyTypeDate",
                "PropertyTypeDecimal",
                "PropertyTypeInteger",
                "PropertyTypeLocalization",
                "PropertyTypeBoolean",
                "PropertyTypeText",
                "PropertyTypeJSONObject",
                "PropertyTypeJSONArray",
                "PropertyTypeMedia",
                "PropertyTypeEnum"
            ]
        },
...
        "parier-server_internal_models.User": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "email": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "realm": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "session_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "parier-server_internal_repository.PropertiesEnumFilter": {
            "type": "object",
            "properties": {
                "language": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "property_type": {
                    "type": "string"
                },
                "search": {
                    "type": "string"
                },
                "sort_by": {
                    "type": "string"
                },
                "sort_dir": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/parier-server_internal_models.User"
                }
            }
        },
        "parier-server_internal_repository.PropertiesTypeFilter": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "place": {
                    "$ref": "#/definitions/parier-server_internal_models.PropertyPlace"
                },
                "search": {
                    "type": "string"
                },
                "sort_by": {
                    "type": "string"
                },
                "sort_dir": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/parier-server_internal_models.User"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OAuth2Keycloak": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "http://localhost:18080/realms/parier/protocol/openid-connect/auth",
            "scopes": {
                "read_access": "Grants read access",
                "write_access": "Grants write access"
            }
        }
    }
}